First version done...
Hobbit rogue, Dracon fighter and Mook psionic are present in the vast majority of parties.
That's because the races are mostly one-class ponies and this has a 1 class and 1 race limit per party.
Trying to relax requirements  (removing them from all roles to promote other heroes) isn't helping.
The race-class identity is fine and organic, the problem is having 9/10 parties have 3/6 fixed heroes.