Fix extreme fighter/rogue/mage/priest/alchemist/psionic bias...
Made humans the "expert race", skipping all common classes.
Made half of the race pool humans, now 50% of heroes are "uncommon" classes.
Sucks to have 2-3 humans most parties but humans are genearlly the most populous race in-lore...
Sucks to have 2 same-gender humans in a party often but beats having a fixed 6/15 party composition with only race swaps 90% of the time.